Table generator by default first handles csv and then html output.
This effectively only changes the order the log messages "Writing x to
file" are displayed.
Output formats themselves are written to in parallel.
Because the csv is usually done before the html and the message order
implies that HTML writing is already done when "Writing csv to file"
is displayed (which is wrong), change the order of output to better fit
the actual behavior.